Add assignment/comparison counting to direct sorting functions
The counting is handled through an optional parameter - I chose this
approach (instead of a return value) so that the user can indicate they do
*not* want counting by leaving the parameter NULL.
Currently, the code for incrementing this counter is scattered throughout
the sorting code. It would be better to extract the counting functionality
into some kind of wrapper around the comparison and assignment functions,
so that the sorting functions themselves don't need to know about it.