Better separate the sorting algorithms from the operation counting
I tried a number of different approaches to separate counting from sorting,
including preprocessor macros, separate functions (with more parameters)
that increment a counter and then call the std utilities, and finally
using an object wrapper that mimics the std utilities as closely as
possible.

The current approach is probably quite inefficient, and may incur a
relevant performance penalty due to the additional objects involved,
however it provides the most similar interface to simply using the std
utilities directly without any counting (hence hiding the counting from the
sorting). As such, I believe it to be the lesser evil for now.